+++ .gitignore
@@ -1,0 +1,5 @@
+/build/
+/bin/
+/.settings/
+/.classpath
+/.project
