+++ src/main/java/com/cryptotronix/device/jwt/DeviceRegistrationUtilities.java
