+++ src/main/java/com/cryptotronix/device/jwt/RegisteredDevice.java
