+++ src/main/java/com/cryptotronix/device/sensor/Sensor.java
