+++ src/main/java/com/cryptotronix/device/sensor/UsedMemorySensor.java
