+++ src/main/resources/device.properties
